fixed the pop ups for adding/deleting a member; in order to test "edit information" for schools, a backend route needs to be added

+++ client/src/pages/HomePage_new.jsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useState, useRef } from 'react';
+import React, {useEffect, useState, useRef} from 'react';
 import axios from 'axios';
-import { useParams, useNavigate } from 'react-router-dom';
+import {useParams, useNavigate} from 'react-router-dom';
 import '../assets/style/HomePage_new.css';
 import Navbar from "../components/Navbar.jsx";
 
 function HomePage_new() {
-    const { id } = useParams();
+    const {id} = useParams();
     const navigate = useNavigate();
     const [schoolId, setSchoolId] = useState(id || null);
     const [schoolName, setSchoolName] = useState('');
@@ -82,13 +82,13 @@ function HomePage_new() {
                 if (schoolId) {
                     response = await axios.get(`http://localhost:4000/schools/${schoolId}`);
                 } else {
-                    response = await axios.get('http://localhost:4000/profile', { withCredentials: true });
-                    const { school, schoolId } = response.data;
+                    response = await axios.get('http://localhost:4000/profile', {withCredentials: true});
+                    const {school, schoolId} = response.data;
                     setSchoolId(schoolId);
                     setSchoolName(school.schoolName);
-                    setSchoolAddress(school.address || '8827 Goldenwood Lake Ct, Boynton Beach FL, 33473');
-                    setSchoolPhone(school.phone || '5619005802');
-                    setSchoolEmail(school.email || 'cathy.t.quan@gmail.com');
+                    // setSchoolAddress(school.address || '8827 Goldenwood Lake Ct, Boynton Beach FL, 33473');
+                    // setSchoolPhone(school.phone || '5619005802');
+                    // setSchoolEmail(school.email || 'cathy.t.quan@gmail.com');
                     setShepInfo(school.SHEP);
                     setGccInfo(school.GCC);
                     setGrades(school.Classes.sort((a, b) => a.className.localeCompare(b.className)));
@@ -96,9 +96,9 @@ function HomePage_new() {
                 if (response.data.school) {
                     const school = response.data.school;
                     setSchoolName(school.schoolName);
-                    setSchoolAddress(school.address || '8827 Goldenwood Lake Ct, Boynton Beach FL, 33473');
-                    setSchoolPhone(school.phone || '5619005802');
-                    setSchoolEmail(school.email || 'cathy.t.quan@gmail.com');
+                    // setSchoolAddress(school.address || '8827 Goldenwood Lake Ct, Boynton Beach FL, 33473');
+                    // setSchoolPhone(school.phone || '5619005802');
+                    // setSchoolEmail(school.email || 'cathy.t.quan@gmail.com');
                     setShepInfo(school.SHEP);
                     setGccInfo(school.GCC);
                     setGrades(school.Classes.sort((a, b) => a.className.localeCompare(b.className)));
@@ -145,9 +145,9 @@ function HomePage_new() {
         if (window.confirm(`Are you sure you want to delete the grade ${selectedGrade}?`)) {
             try {
                 await axios.delete(`http://localhost:4000/schools/${schoolId}/deleteClass`, {
-                    data: { className: selectedGrade }
+                    data: {className: selectedGrade}
                 });
-                setGrades(grades.filter(grade => grade.className !== selectedGrade).sort((a, b) => a.className.localeCompare(b.className))); 
+                setGrades(grades.filter(grade => grade.className !== selectedGrade).sort((a, b) => a.className.localeCompare(b.className)));
                 setShowDeleteGradePopup(false);
             } catch (error) {
                 console.error('There was an error deleting the grade!', error);
@@ -166,10 +166,10 @@ function HomePage_new() {
             const response = await axios.post(`http://localhost:4000/register`, {
                 name: newMemberName,
                 role: newMemberRole,
-                email: newMemberEmail,
+                email: newMemberEmail.toLowerCase(),
                 phone: newMemberPhone,
                 password: newMemberName,
-                school: id
+                school: schoolId
             });
 
             setMembers([...members, response.data]);
@@ -188,13 +188,22 @@ function HomePage_new() {
         } catch (error) {
             console.error('There was an error adding the member!', error);
         }
+        if (error.response && error.response.data && error.response.data.error) {
+            alert(error.response.data.error);
+        } else {
+            alert('There was an error adding the member. Please try again.');
+        }
     };
 
     const handleDeleteMember = async (e) => {
         e.preventDefault();
+        const confirmed = window.confirm(`Are you sure you want to delete the member ${memberToDelete}? This action cannot be undone.`);
+        if (!confirmed) {
+            return;
+        }
         try {
             const response = await axios.delete(`http://localhost:4000/deleteMember`, {
-                data: { name: memberToDelete, school: id }
+                data: {name: memberToDelete, school: schoolId}
             });
 
             if (response.data.message === 'Member deleted successfully.') {
@@ -204,6 +213,10 @@ function HomePage_new() {
                     setShepInfo(null);
                 } else if (gccInfo && gccInfo.name === memberToDelete) {
                     setGccInfo(null);
+                }
+                if (user.name === memberToDelete) {
+                    setUser(null);  // Clear the user context
+                    navigate('/login');  // Redirect to login page
                 }
 
                 setShowDeleteMemberPopup(false);
@@ -283,7 +296,7 @@ function HomePage_new() {
                 <h1>{schoolName}</h1>
             </div>
             <div className="main-content">
-                <div className="school-info" style={{ width: `${leftWidth}%` }}>
+                <div className="school-info" style={{width: `${leftWidth}%`}}>
                     <div className="school-info-container">
                         <div className="basic-school-info">
                             <p>{schoolAddress}</p>
@@ -330,7 +343,7 @@ function HomePage_new() {
                     </div>
                 </div>
                 <div className="divider" ref={dividerRef}></div>
-                <div className={`grade-list ${getGridClass()}`} style={{ width: `${100 - leftWidth}%` }}>
+                <div className={`grade-list ${getGridClass()}`} style={{width: `${100 - leftWidth}%`}}>
                     {grades.map(grade => (
                         <button key={grade._id} onClick={() => handleGradeClick(grade._id)}>{grade.className}</button>
                     ))}
@@ -362,11 +375,11 @@ function HomePage_new() {
                             </label>
                             <label>
                                 Teacher Email:
-                                <input 
-                                    type="email" 
-                                    value={teacherEmail} 
-                                    onChange={(e) => setTeacherEmail(e.target.value)} 
-                                    required 
+                                <input
+                                    type="email"
+                                    value={teacherEmail}
+                                    onChange={(e) => setTeacherEmail(e.target.value)}
+                                    required
                                 />
                             </label>
                             <div className="popup-buttons">
